django_project/requirements.txt
Django==3.0.6
django-crispy-forms==1.9.1
django-extensions==2.2.9
django-fsm==2.7.0
django-getpaid==2.2.1
django-getpaid-payu==0.2.0
django-model-utils==4.0.0
Pillow==7.1.2
